// apps/frontend/src/renderer/PageShell.tsx
import React, { Suspense, useEffect } from "react";
import { PageContextProvider } from "./PageContextProvider";
import type { PageContext } from "./types";
import { ThemeProvider } from "./ThemeProvider";

interface Props {
  readonly pageContext: PageContext;
  readonly children: React.ReactNode;
}

export function PageShell(props: Props) {
  const { pageContext, children } = props;

  useEffect(() => {
    const documentProps =
      pageContext.documentProps ?? pageContext.exports.documentProps;

    document.title = documentProps?.title ?? document.title;
  }, [pageContext.pageProps]);

  return (
    <React.StrictMode>
      <PageContextProvider pageContext={pageContext}>
        <ThemeProvider>
          <Suspense fallback={<p>Loading...</p>}>
            {(pageContext.hasLayout ?? true) === true ? (
              <PageLayout pageContext={pageContext}>{children}</PageLayout>
            ) : (
              children
            )}
          </Suspense>
        </ThemeProvider>
      </PageContextProvider>
    </React.StrictMode>
  );
}

function PageLayout(props: Props) {
  const { children } = props;

  return <>{children}</>;
}
